fix cropping for native flavor
