+++ gradle.properties
@@ -17,8 +17,6 @@
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
 
-VERSION_NAME=2.2.9
-VERSION_CODE=28
 GROUP=com.yalantis
 
 POM_DESCRIPTION=Android Library for cropping images
