+++ ucrop/src/main/java/com/yalantis/ucrop/backend/UCropBackendType.java
@@ -1,0 +1,12 @@
+package com.yalantis.ucrop.backend;
+
+public enum UCropBackendType {
+    NATIVE("NATIVE"),
+    DEFAULT("NON_NATIVE");
+
+    public final String label;
+
+    private UCropBackendType(String label) {
+        this.label = label;
+    }
+}
